+++ admin/metabox/ui/components/Modal/Modal.js
@@ -4,6 +4,7 @@ import { v4 as uuid } from 'uuid';
 
 import ErrorMessage from 'metabox/components/ErrorMessage/ErrorMessage';
 import Spinner from 'metabox/components/Spinner/Spinner';
+
 import { AdminContext } from 'metabox/context/adminContext';
 import { selectForm, selectTitle } from 'metabox/utils/select-form';
 import { updatePost } from 'metabox/utils/update-post';